src/api/requirements.txt
deta
fastapi
python-dotenv
uvicorn[standard]==0.17.6